+++ components/util/Error.tsx
@@ -1,4 +1,4 @@
-import styles from "../../styles/components/Error.module.css";
+import styles from "../../styles/components/util/Error.module.css";
 import React from "react";
 
 // Error text
